+++ CHANGELOG.md
@@ -109,6 +109,8 @@ Minor / Fixes:
   * launch targets with the tag 'autorun' will be launched on
 	  startup
 
+  * display orientation options are now explicit +- 90
+
 # 0.4.0
   * Display region sharing now supports force-pushing into clients
     that can handle input segments.